Add main node and FSM skeleton.
The main node is made by a class, whose run method implements the main
loop. In this method, the FSM is updated on each cycle so far. It also
contains a set of topics the node subscribes to, but they are only a
sketch of how it should be in the future.

The FSM contains the update method, where the state of the FSM will be
updated. Right now it only contains the states considered for the drone
controller. The considered state variables are included as well. They
define how the system will behave i.e. how the FSM wil evolve.