Add test case with all the expected transitions in the FSM.
The added test case checks that the FSM evolves through the
different states as expected, taking as reference the FSM diagram.

There is a second test case where different behaviours of the FSM
are tested. To begin with, a test for checking if the waypoints are
correctly updated was added. Almost any unexpected behaviours are
included in the test case so far.