Add callbacks for the position and acknowledge messages
The position of the drone is registered every time a value is published to the topic.
Also, every time that the gcs_master sends a command, an acknoledge message is obtained
back. Further work has to deal with the situation where the acknowledge is different from
0 (True).

The print() functions are substituted by log messages. The condition for evolving from the
taking_off state is changed, so now it is checked if the drone is over the altitude set
point.