Finish functions for the flying, calc_path, and upload_mission states
The former flying state is divided into 3, so the functionality is
more complete. The corresponding flags are updated after the state
transitions, and a timer is used for setting them again after certain
time.

The landed state is removed. It is the start state (maybe this change
needs to be reverted).

Unit tests are created for testing the transition until the flying
state, including going through calculate_path and upload_mission.