Correct bugs at take-off and give meaningful names to flags.
The flags in the UPLOAD_MISSION and START_MISSION states were confusing,
due to the similiraties between the states names. Now, the flags have the
same name as the state that sets them.

The communication with the drone needs to be working before evolving to
the arm state.

The FSM was never evolving to the START_MISSION state. The transition
has been fixed.

Plot the warning text when an acknowledge is not successful.