Solve bugs and add functionalities, so a Gazebo simulation can be implemented
A new subscriber is added, for getting the destination position. The drone positionn is
updated and saved continuously (only X and Y coordinates).

The transition from flying to landing is corrected. It waits until the drone is doing the
last waypoint, and it is very close to the destination (A threshold is used).

The transition to calculate_path is temporary commented