Disarm the drone when it reaches back the start state.
A new flag is added to the FSM for that purpose. Since
the arm and disarm commands share the same topic, when
the command is acknowledged, it has to be checked as
well in which state the FSM is, and set the arm variable
to true only when it is not in the FSM state.