Land drone by uploading mission (#66). Fix bug (#31)
In order to avoid a situation where the program is restarted
during flight and the drone gets disarmed, now the landing is
done by sending a single item mission to the drone with the
land command.

Now the GCS FSM waits as well for the GPS to send values before
starting execution, so there are no errors afterwards (#31).

The simulation path is now a linear path from A to B. The distance
to destination threshold is corrected, as it was taking into
account the wrong distance.

The mission is cleared from the drone when it lands and goes back
to the start state