Add the correct information about heading and speed to UTM
The next waypoint heading is obtained from the path planner. The
speed of the next waypoint is assumed to always be 5 m/s.

The current speed is obtained from the drone GPS. It does not work
with the Gazebo simulation, so it has to be tested with the real
drone.